Refactor and bugfix
Refactor updateDisplay() to accept an argument for what to display on screen. Makes function more dynamic and adaptable to different situations

Add handleDivideByZero() to properly handle divison by zero and conditionally call this function in the places where operate() is used. This implementation prevents users from operating on either javascript's incorrect 'infinity' value or on the error message.

Fix zero check in typeToCurrentValue() to work with the user operating with a single '0'.